+++ src/Microsoft.ServiceFabric.Actors.KVSToRCMigration/Controllers/RcMigrationController.cs
@@ -47,38 +47,5 @@ namespace Microsoft.ServiceFabric.Actors.KVSToRCMigration.Controllers
         {
             throw new NotImplementedException();
         }
-
-        /// <summary>
-        /// Starts the Downtime phase on the current partition. In the downtime phase all the actor calls are actively rejected with MigrationException.
-        /// </summary>
-        /// <param name="cancellationToken">Token to signal cancellation on the asynchronous operation</param>
-        /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
-        [HttpPut("StartDowntime")]
-        public async Task StartDowntimeAsync(CancellationToken cancellationToken)
-        {
-            await this.MigrationOrchestrator.StartDowntimeAsync(cancellationToken);
-        }
-
-        /// <summary>
-        /// Aborts the Actor state migration on the current partition.
-        /// </summary>
-        /// <param name="cancellationToken">Token to signal cancellation on the asynchronous operation</param>
-        /// <returns>A <see cref="Task"/> representing the result of the asynchronous operation.</returns>
-        [HttpPut("AbortMigration")]
-        public async Task AbortMigrationAsync(CancellationToken cancellationToken)
-        {
-            await this.MigrationOrchestrator.AbortMigrationAsync(cancellationToken);
-        }
-
-        /// <summary>
-        /// Starts the Actor state migration on the current partition.
-        /// </summary>
-        /// <param name="cancellationToken">Token to signal cancellation on the asynchronous operation</param>
-        /// <returns>A <see cref="Task"/> representing the result of the asynchronous operation.</returns>
-        [HttpPut("StartMigration")]
-        public async Task StartMigrationAsync(CancellationToken cancellationToken)
-        {
-            await this.MigrationOrchestrator.StartMigrationAsync(cancellationToken);
-        }
     }
 }
